Split checklist popover pointer tracking and summary modifier into own files
Keeps SidebarWorkspaceChecklistPopover.swift and
SidebarWorkspaceChecklistView.swift under the 500-line tracking
threshold of the Swift file-length budget gate. Pure moves:
PopoverPointerTracker + ChecklistPopoverRowFramesKey to
ChecklistPopoverPointerTracking.swift, ChecklistSummaryPopoverModifier
to ChecklistSummaryPopoverModifier.swift (private -> internal).
